style(fst-dict): move relevant tests from full_dictionary.rs to mod.rs
